The algorithm appears to work well and the code has been significantly cleaned and improved exspecially state2State1 and state12State functions

+++ Environment/fmu-pvs/misraC/Environment.h
@@ -158,8 +158,8 @@ typedef struct {
  */
 typedef struct {
     Cell map[10][10]; //-- real matrix
-    int32_t x[4];
-    int32_t y[4];
+    float64_t x[4];
+    float64_t y[4];
     Cell* neighbourhood;
     int32_t k;
 } State1;
